-add select method for main activity

+++ src/andy/study/dailyrecord/dao/DataManager.java
@@ -186,9 +186,9 @@ public class DataManager {
 					String value = rawQuery.getString(rawQuery.getColumnIndex(name));
 					if (currentField.getType() == int.class || currentField.getType() == Integer.class) {
 						currentField.setInt(record, Integer.parseInt(value == null ? "0":value));
-					} else if (currentField.getType() == float.class || currentField.getType() == Double.class) {
+					} else if (currentField.getType() == Double.class || currentField.getType() == double.class) {
 						DecimalFormat format = new DecimalFormat("#.00");
-						currentField.setFloat(record, Float.parseFloat(format.format(value == null ? "0":value)));
+						currentField.setDouble(record, Double.parseDouble(format.format(value == null ? "0":value)));
 					} else if (currentField.getType() == boolean.class || currentField.getType() == Boolean.class) {
 						currentField.setBoolean(record, Boolean.parseBoolean(value == null ? "false":value));
 					} else {
